fix err int to double of Order

+++ lib/data/models/order/info_production_model.dart
@@ -59,7 +59,7 @@ class InfoProduction {
       songB_Replace: json['songB_Replace'] ?? "",
       songC_Replace: json['songC_Replace'] ?? "",
       songE2_Replace: json['songE2_Replace'] ?? "",
-      sizePaper: (json['sizePaper'] ?? 0).toDouble(),
+      sizePaper: json['sizePaper'] ?? 0.0,
       quantity: json['quantity'] ?? 0,
       instructSpecial: json['instructSpecial'] ?? "",
       numberChild: json['numberChild'] ?? 0,
